Move paths to its own package cmd/agent/common/path - Remove dependency cmd/agent/common from com/core/config (to break circle) - Code for common_windows is split
