+++ cmd/agent/gui/checks.go
@@ -22,6 +22,7 @@ import (
 	yaml "gopkg.in/yaml.v2"
 
 	"github.com/DataDog/datadog-agent/cmd/agent/common"
+	"github.com/DataDog/datadog-agent/cmd/agent/common/path"
 	"github.com/DataDog/datadog-agent/pkg/autodiscovery/integration"
 	"github.com/DataDog/datadog-agent/pkg/collector"
 	"github.com/DataDog/datadog-agent/pkg/collector/check"
@@ -32,14 +33,14 @@ import (
 
 var (
 	configPaths = []string{
-		config.Datadog.GetString("confd_path"),        // Custom checks
-		filepath.Join(common.GetDistPath(), "conf.d"), // Default check configs
+		config.Datadog.GetString("confd_path"),      // Custom checks
+		filepath.Join(path.GetDistPath(), "conf.d"), // Default check configs
 	}
 
 	checkPaths = []string{
-		filepath.Join(common.GetDistPath(), "checks.d"), // Custom checks
-		config.Datadog.GetString("additional_checksd"),  // Custom checks
-		common.PyChecksPath,                             // Integrations-core checks
+		filepath.Join(path.GetDistPath(), "checks.d"),  // Custom checks
+		config.Datadog.GetString("additional_checksd"), // Custom checks
+		path.PyChecksPath, // Integrations-core checks
 	}
 )
 
@@ -234,10 +235,10 @@ func setCheckConfigFile(w http.ResponseWriter, r *http.Request) {
 
 	if checkFolder != "" {
 		checkConfFolderPath = filepath.Join(config.Datadog.GetString("confd_path"), checkFolder)
-		defaultCheckConfFolderPath = filepath.Join(common.GetDistPath(), "conf.d", checkFolder)
+		defaultCheckConfFolderPath = filepath.Join(path.GetDistPath(), "conf.d", checkFolder)
 	} else {
 		checkConfFolderPath = config.Datadog.GetString("confd_path")
-		defaultCheckConfFolderPath = filepath.Join(common.GetDistPath(), "conf.d")
+		defaultCheckConfFolderPath = filepath.Join(path.GetDistPath(), "conf.d")
 	}
 
 	if r.Method == "POST" {
