+++ cmd/agent/subcommands/jmx/command_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/DataDog/datadog-agent/cmd/agent/command"
-	"github.com/DataDog/datadog-agent/cmd/agent/common"
+	"github.com/DataDog/datadog-agent/cmd/agent/common/path"
 	"github.com/DataDog/datadog-agent/comp/core"
 	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
 )
@@ -64,7 +64,7 @@ func TestCollectCommand(t *testing.T) {
 				require.Equal(t, "collect", cliParams.command)
 				require.Equal(t, "debug", cliParams.jmxLogLevel)      // overrides --log-level
 				require.Equal(t, "debug", coreParams.LogLevelFn(nil)) // overrides --log-level
-				require.True(t, strings.HasPrefix(cliParams.logFile, common.DefaultJMXFlareDirectory))
+				require.True(t, strings.HasPrefix(cliParams.logFile, path.DefaultJMXFlareDirectory))
 				require.Equal(t, cliParams.logFile, coreParams.LogFileFn(nil))
 				require.Equal(t, "CORE", coreParams.LoggerName())
 				require.Equal(t, true, coreParams.ConfigLoadSecrets())
