+++ cmd/security-agent/command/command.go
@@ -7,11 +7,12 @@ package command
 
 import (
 	"fmt"
-	"github.com/fatih/color"
-	"github.com/spf13/cobra"
 	"path"
 
-	"github.com/DataDog/datadog-agent/cmd/agent/common"
+	"github.com/fatih/color"
+	"github.com/spf13/cobra"
+
+	commonpath "github.com/DataDog/datadog-agent/cmd/agent/common/path"
 	"github.com/DataDog/datadog-agent/cmd/security-agent/flags"
 )
 
@@ -24,8 +25,8 @@ type SubcommandFactory func(globalParams *GlobalParams) []*cobra.Command
 const LoggerName = "SECURITY"
 
 var defaultSecurityAgentConfigFilePaths = []string{
-	path.Join(common.DefaultConfPath, "datadog.yaml"),
-	path.Join(common.DefaultConfPath, "security-agent.yaml"),
+	path.Join(commonpath.DefaultConfPath, "datadog.yaml"),
+	path.Join(commonpath.DefaultConfPath, "security-agent.yaml"),
 }
 
 // MakeCommand makes the top-level Cobra command for this command.
